fix: thread collapse_versioned_symbols through extracted post-processing; split cli_buildsource under cap
After main was merged into this branch, two regressions surfaced from the
interaction of that merge with the complexity refactor:

1. `collapse_versioned_symbols` (added on main) was not forwarded by the
   extracted `_run_post_processing` helper, so `DEFAULT_PIPELINE.run` kept its
   default `False` — versioned rename churn stayed as func_removed/func_added
   and the verdict stayed BREAKING. Thread the flag through
   `_run_post_processing` and `compare()`'s call site. Fixes
   tests/test_versioned_symbol_scheme.py::test_collapse_preset_reclassifies_versioned_pairs.
   (Credit: codex PR review caught this.)

2. The merge pushed `cli_buildsource.py` to 2005 lines, over the 2000-line
   hard cap. Extract the `merge` command's helpers
   (`_merge_load_snapshots`, `_merge_pick_base`, `_merge_fold_packs`,
   `_merge_handle_conflicts`, `_merge_attach_combined`, `_merge_print_summary`,
   `_exported_symbols_from_snapshot`) into a new sibling module
   `cli_buildsource_helpers.py` (no import cycle). cli_buildsource.py → 1882 lines.

Verified: full fast suite 10,188 passed / 0 failed; mypy 0 errors;
ai-readiness 0 errors; ruff clean.
